connector/requirements.txt
anthropic>=0.40.0
certifi>=2024.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
rapidfuzz>=3.9.0
